move spreadsheetid logic to a hook and get lookup brick working
